importccl: move table creation helpers to own file
Release note: none.
